requirements.txt
# ================================================
# JOBSEARCHAI PLATFORM - DEPENDENCIAS
# Actualizado: Noviembre 2025
# Version: 4.0.0 (Busqueda de Empleo con IA)
# ================================================

# ------------------------------------------------
# Django Core
# ------------------------------------------------
Django==5.1.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.10.0

# ------------------------------------------------
# Database
# ------------------------------------------------
psycopg2-binary==2.9.9  # PostgreSQL adapter
dj-database-url==2.2.0   # Database URL parsing

# ------------------------------------------------
# Authentication & Security
# ------------------------------------------------
django-allauth==0.63.3   # Enhanced authentication
argon2-cffi==23.1.0      # Strong password hashing

# ------------------------------------------------
# Agent_IA Core - LangChain
# ------------------------------------------------
langchain-core>=0.3.72,<1.0.0
langchain>=0.3.27,<0.4.0
langchain-community>=0.3.0,<0.4.0

# LangChain Providers
langchain-openai>=0.3.35           # OpenAI GPT-4
langchain-google-genai>=2.0.10     # Google Gemini
langchain-ollama>=0.2.0,<1.0.0     # Ollama (Local)

# ------------------------------------------------
# LLM APIs
# ------------------------------------------------
openai>=1.61.0          # OpenAI API client
google-generativeai>=0.8.3  # Google Gemini API
tiktoken>=0.8.0         # OpenAI tokenizer
httpx>=0.28.0           # Required by Ollama
ollama>=0.6.0           # Ollama Python SDK

# ------------------------------------------------
# Google Custom Search API
# ------------------------------------------------
google-api-python-client>=2.100.0

# ------------------------------------------------
# PDF Processing
# ------------------------------------------------
PyPDF2>=3.0.0           # Extract text from PDF CVs
# pdfminer.six>=20221105  # Alternative PDF library

# ------------------------------------------------
# XML Processing & Data Validation
# ------------------------------------------------
lxml==5.3.0
requests==2.32.3
pydantic==2.12.2
pydantic-settings==2.7.0
python-dateutil==2.9.0.post0

# ------------------------------------------------
# Media & File Processing
# ------------------------------------------------
Pillow==11.3.0

# ------------------------------------------------
# Utilities
# ------------------------------------------------
python-dotenv==1.0.1
tqdm==4.67.1
PyYAML==6.0.2
markdown>=3.4.1         # Markdown to HTML for chat
Pygments>=2.19.0        # Syntax highlighting

# ------------------------------------------------
# Monitoring & Logging
# ------------------------------------------------
sentry-sdk==2.19.2
django-debug-toolbar==4.4.6

# ------------------------------------------------
# Testing
# ------------------------------------------------
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
faker==33.1.0

# ------------------------------------------------
# Code Quality
# ------------------------------------------------
black==24.10.0
isort==5.13.2
flake8==7.1.1

# ------------------------------------------------
# Production
# ------------------------------------------------
django-storages==1.14.4  # S3 storage
boto3==1.35.77           # AWS SDK

# ------------------------------------------------
# Additional Dependencies
# ------------------------------------------------
orjson==3.11.3
httpcore==1.0.9
typing-extensions==4.15.0
annotated-types==0.7.0

# ================================================
# NOTAS DE INSTALACION
# ================================================
#
# Para instalar todas las dependencias:
#   pip install -r requirements.txt
#
# Para instalar solo las dependencias core:
#   pip install Django langchain langchain-ollama google-api-python-client PyPDF2
#
# OLLAMA:
#   Ollama NO requiere instalacion via pip.
#   Descargar desde: https://ollama.com
#
# GOOGLE CUSTOM SEARCH:
#   Configurar API key y Engine ID en perfil de usuario
#   Mas info: https://developers.google.com/custom-search
#
# COMPATIBILIDAD:
#   Python: 3.10, 3.11, 3.12
#   Sistema Operativo: Windows, Linux, macOS
#
# ================================================
